fix codes based on reviewer's request
